+++ src/index.ts
@@ -4,9 +4,12 @@ export * from './bedita-api-client';
 export { default as AuthInterceptor } from './interceptors/auth-interceptor';
 export { default as ContentTypeInterceptor } from './interceptors/content-type-interceptor';
 export { default as FormatUserInterceptor } from './interceptors/format-user.interceptor';
-export { default as MapIncludedInterceptor } from './interceptors/map-included-interceptor';
+export * from './interceptors/map-included-interceptor';
 export { default as RefreshAuthInterceptor } from './interceptors/refresh-auth-interceptor'
 export { default as RequestInterceptor, RequestInterceptorInterface } from './interceptors/request-interceptor';
 export { default as ResponseInterceptor, ResponseInterceptorInterface } from './interceptors/response-interceptor';
 export { default as StorageService } from './services/storage-service';
-/* c8 ignore start */
+export { default as StorageAdapterInterface } from './services/adapters/storage-adapter-interface';
+export { default as LocalStorageAdapter } from './services/adapters/local-storage-adapter';
+export { default as MemoryStorageAdapter } from './services/adapters/memory-storage-adapter';
+/* c8 ignore stop */